Updates a large swathe of the model to use case class derived json formats
