+++ app/model/Section.scala
@@ -2,8 +2,9 @@ package model
 
 import com.amazonaws.services.dynamodbv2.document.Item
 import helpers.XmlHelpers._
+import org.cvogt.play.json.Jsonx
+import org.cvogt.play.json.implicits.optionWithNull
 import play.api.Logger
-import play.api.libs.functional.syntax._
 import play.api.libs.json.{JsValue, Json, JsPath, Format}
 import com.gu.tagmanagement.{Section => ThriftSection}
 
@@ -53,18 +54,7 @@ case class Section(
 
 object Section {
 
-  implicit val sectionFormat: Format[Section] = (
-      (JsPath \ "id").format[Long] and
-      (JsPath \ "sectionTagId").format[Long] and
-      (JsPath \ "name").format[String] and
-      (JsPath \ "path").format[String] and
-      (JsPath \ "wordsForUrl").format[String] and
-      (JsPath \ "pageId").format[Long] and
-      (JsPath \ "editions").formatNullable[Map[String, EditionalisedPage]].inmap[Map[String, EditionalisedPage]](_.getOrElse(Map()), Some(_)) and
-      (JsPath \ "discriminator").formatNullable[String] and
-      (JsPath \ "isMicrosite").format[Boolean] and
-      (JsPath \ "activeSponsorships").formatNullable[List[Long]].inmap[List[Long]](_.getOrElse(Nil), Some(_))
-    )(Section.apply, unlift(Section.unapply))
+  implicit val sectionFormat = Jsonx.formatCaseClassUseDefaults[Section]
 
   def fromItem(item: Item) = try{
     Json.parse(item.toJSON).as[Section]
