+++ app/model/TrackingInformation.scala
@@ -1,7 +1,7 @@
 package model
 
 import play.api.libs.json._
-import play.api.libs.functional.syntax._
+import org.cvogt.play.json.Jsonx
 import com.gu.tagmanagement.{TrackingInformation => ThriftTrackingInformation}
 
 case class TrackingInformation(trackingType: String) {
@@ -17,9 +17,7 @@ case class TrackingInformation(trackingType: String) {
 
 object TrackingInformation {
 
-  implicit val trackingReads: Reads[TrackingInformation] = (JsPath \ "trackingType").read[String].map(TrackingInformation(_))
-
-  implicit val trackingWrite: Writes[TrackingInformation] = (JsPath \ "trackingType").write[String].contramap(_.trackingType)
+  implicit val trackingFormat = Jsonx.formatCaseClass[TrackingInformation]
 
   def apply(thriftTrackingInformation: ThriftTrackingInformation): TrackingInformation =
     TrackingInformation(
